Renamed TheRoom to MetaMansion
* did a couple of tweaks after feedback from @peth-yursick.
* MG crystal icon now opens new tab when in the Mansion.
