+++ src/pages/metaMansion.module.scss
@@ -5,9 +5,8 @@
  */
 
 .theRoom {
-  height: calc(100vh - 75px);
-  max-height: calc(100vh - 75px);
-  overflow: hidden;
+  height: calc(100vh);
+  max-height: calc(100vh - 60px);
   width: 100%;
   border-width: 0;
 }
